requirements.txt
discord-py-interactions[speedup]==5.10.0
tansy==0.8.0
python-dotenv==1.0.0
humanize==4.7.0
orjson==3.9.4
redis[hiredis]==4.6.0
types-redis>=4.4,<4.7
prisma==0.10.0
argon2-cffi==23.1.0
pycryptodome==3.19.0
uvloop==0.17.0; platform_system == "Linux"